+++ CHANGELOG.md
@@ -11,6 +11,7 @@ This document contains what changed in each release of Fisher.
  * Add the `max_threads` field to `GET /health`
  * Add the `label`, `milestone`, `organization`, `project_card`,
    `project_column`, `project`, `pull_request_review`, `team` GitHub events
+ * Add the ability to provide extra environment variables with the `-e` flag
 
 * **Changes and improvements:**